Tidy temp E2E test
